Reorganize solution to fix MonoGame patches not being applied
