--- SDV.Installer/SDV.Installer.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <AssemblyName>SDV.Installer</AssemblyName>
-    <RootNamespace>SDV.Installer</RootNamespace>
-    <Version>1.1.1</Version>
-    <TargetFramework>net452</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="MonoMod" Version="21.3.1.1" />
-    <PackageReference Include="MonoMod.RuntimeDetour" Version="21.3.1.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\StardewValley.Patches.mm\StardewValley.Patches.mm.csproj" />
-  </ItemGroup>
-
-  <Target Name="CopyGameFiles" AfterTargets="AfterBuild">
-    <ItemGroup>
-      <GameFiles Include="..\libs\CopyToGameFolder\*.dll" />
-    </ItemGroup>
-
-    <Copy SourceFiles="..\libs\CorFlags.exe" DestinationFolder="$(OutDir)\libs" />
-    <Copy SourceFiles="@(GameFiles)" DestinationFolder="$(OutDir)\libs\CopyToGameFolder" />
-  </Target>
-</Project>
