+++ Stardew64Installer/Stardew64Installer.csproj
@@ -1,0 +1,28 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\common.targets" />
+
+  <PropertyGroup>
+    <AssemblyName>Stardew64Installer</AssemblyName>
+    <RootNamespace>Stardew64Installer</RootNamespace>
+    <OutputType>Exe</OutputType>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="MonoMod" Version="21.3.1.1" />
+    <PackageReference Include="MonoMod.RuntimeDetour" Version="21.3.1.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MonoGame.Framework.Patches.mm\MonoGame.Framework.Patches.mm.csproj" />
+    <ProjectReference Include="..\StardewValley.Patches.mm\StardewValley.Patches.mm.csproj" />
+  </ItemGroup>
+
+  <Target Name="CopyGameFiles" AfterTargets="AfterBuild">
+    <ItemGroup>
+      <GameFiles Include="..\libs\CopyToGameFolder\*.dll" />
+    </ItemGroup>
+
+    <Copy SourceFiles="..\libs\CorFlags.exe" DestinationFolder="$(OutDir)\libs" />
+    <Copy SourceFiles="@(GameFiles)" DestinationFolder="$(OutDir)\libs\CopyToGameFolder" />
+  </Target>
+</Project>
