--- StardewValley.Patches.mm/Patches/StardewValley/patch_KeyboardDispatcher.cs
@@ -1,41 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Reflection;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Input;
-using MonoMod;
-using MonoMod.Utils;
-using StardewValley.Patches.mm.Framework;
-
-// ReSharper disable once CheckNamespace
-namespace StardewValley
-{
-    // ReSharper disable once ArrangeTypeModifiers
-    /// <summary>A MonoMod patch that reimplements the <see cref="KeyboardDispatcher"/> constructor to remove a Linux-only check around the <see cref="GameWindow.TextInput"/> event set, ensures that <see cref="KeyboardInput.Initialize"/> is called, and removes the unneeded <see cref="KeyboardInput.CharEntered"/> and <see cref="KeyboardInput.KeyDown"/> events.</summary>
-    [SuppressMessage("Style", "IDE1006:Naming Styles")]
-    class patch_KeyboardDispatcher : KeyboardDispatcher
-    {
-        [MonoModConstructor]
-        [SuppressMessage("Style", "IDE1006:Naming Styles")]
-        public void ctor(GameWindow window)
-        {
-            _commandInputs = new List<char>();
-            _keysDown = new List<Keys>();
-            _charsEntered = new List<char>();
-            _window = window;
-
-            new DynData<KeyboardDispatcher>(this).Set("pastedResult", "");
-
-            // https://stackoverflow.com/questions/1121441/addeventhandler-using-reflection
-            // https://stackoverflow.com/questions/11120401/creating-delegate-from-methodinfo
-            EventInfo windowTextInput = PatchHelper.RequireEvent(typeof(GameWindow), "TextInput");
-            Delegate handler = Delegate.CreateDelegate(windowTextInput.EventHandlerType, this, "Event_TextInput");
-            windowTextInput.AddEventHandler(window, handler);
-
-            KeyboardInput.Initialize(window);
-        }
-
-        public patch_KeyboardDispatcher(GameWindow window) : base(window) { }
-    }
-}
