+++ common.targets
@@ -1,0 +1,6 @@
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <Version>1.1.1</Version>
+    <TargetFramework>net452</TargetFramework>
+  </PropertyGroup>
+</Project>
